move lightning module into models module
